refractor: move undistorted functionality from camera-reader  over to lane_following.py
