+++ Exercise 3/exercise-3/.dtproject
@@ -1,0 +1,3 @@
+VERSION=0.0.0
+TYPE=template-ros
+TYPE_VERSION=3
